Fix material in test; fix wrong synchronisation

+++ main/src/main/kotlin/pl/edu/agh/firevox/synchroniser/IterationSynchroniser.kt
@@ -28,7 +28,7 @@ class IterationSynchroniser(
         val iteration = countersRepository.findByIdOrNull(CounterId.CURRENT_ITERATION)?.count!!
         SynchroniserImpl.log.info("Running synchronisation $iteration")
         synchroniserImpl.verifyIterationFinish(iteration)
-        synchroniserImpl.synchroniseRadiationResults(iteration)
+        synchroniserImpl.synchroniseRadiationResults(iteration.toInt())
         synchroniserImpl.resetCounters(iteration)
         verifyJobFinish(iteration)
     }
