+++ worker/src/test/kotlin/pl/edu/agh/firevox/worker/physics/RadiationExecutionTest.kt
@@ -151,7 +151,7 @@ class RadiationExecutionTest(
                 log.info("Finished conduction")
                 planes.parallelStream().forEach { k -> radiationCalculator.calculateWithVoxelsFilled(k, i) }
                 log.info("Finished calculations")
-                synchroniserImpl.synchroniseRadiationResults(i.toLong())
+                synchroniserImpl.synchroniseRadiationResults(i)
                 log.info("Finished synchronisation")
                 countersRepository.increment(CounterId.CURRENT_ITERATION)
                 log.info("Finished increment")
